#42, 44



See merge request !19

+++ k_1c_upload/ParsingModel.php
@@ -484,6 +484,7 @@ class ParsingModel {
             foreach($lines as $line) {
                 $elementXmlID = $line->obj_id;
                 $type = $line->obj_type;
+                $show_in_object = $line->show_in_object;
 
                 if($type == 'brands') {
                     $iblock_id = self::BRAND_IBLOCK_ID;
@@ -607,7 +608,7 @@ class ParsingModel {
                     $param[] = array('VALUE' => $brand['id_b']);
                 }
                 if ($this->addDoc($fileXmlID, $file_obj, $name, $param, $del, $this->classTrans[$type_index],
-                    $access_level, $actual ? "Y" : "N")) {
+                    $access_level, $actual ? "Y" : "N", $show_in_object ? "Y" : "N")) {
                     $answer['StringTehDoc'][] = $fileXmlID;
                 }
             }
@@ -625,7 +626,7 @@ class ParsingModel {
                                 );
                             }
                             $id_doc_dynamic = $this->addDoc($fileXmlID, $file_obj, $name, $arFile, $del,
-                                $this->classTrans[$type_index], $access_level);
+                                $this->classTrans[$type_index], $access_level, $actual ? "Y" : "N", $show_in_object ? "Y" : "N");
 
                             $id_doc_dynamic = array(array('VALUE' => $id_doc_dynamic));
                             if($this->setPropertyValue($product['id_b'], "docs", $id_doc_dynamic)) {
@@ -776,7 +777,7 @@ class ParsingModel {
         return $el->SetPropertyValueCode($objId, $property_code, $values);
     }
 
-    function addDoc($xmlId, $file, $name, $brand, $del, $type, $access_level = 1, $active = "Y") {
+    function addDoc($xmlId, $file, $name, $brand, $del, $type, $access_level = 1, $active = "Y", $show_in_object = "N") {
 
         $arLoadArray = array(
             "IBLOCK_SECTION_ID" => false,
@@ -784,13 +785,17 @@ class ParsingModel {
             "XML_ID" => $xmlId,
             "PROPERTY_VALUES" => array(
                 'type' => $type,
-                'file' => $file,
                 'brand' => $brand,
-                'access_level' => $access_level
+                'access_level' => $access_level,
+                'show_in_object' => $show_in_object
             ),
             "NAME" => $name,
             "ACTIVE" => $active,
         );
+
+        if (!empty($file)) {
+            $arLoadArray["PROPERTY_VALUES"]["file"] = $file;
+        }
 
         $el = new CIBlockElement;
         $item = $this->findEl(array('IBLOCK_ID' => self::DOCUMENTATION_IBLOCK_ID, "=XML_ID" => $xmlId));
@@ -808,7 +813,7 @@ class ParsingModel {
                 if($el->Update($item['ID'], $arLoadArray)) {
                     echo 'Документация обновлена: ['.$item['ID'].'] '.$xmlId.'<br>';
                     $id_r = $item['ID'];
-                    if($this->answer) {
+                    if($this->answer && !empty($file)) {
                         unlink($file['tmp_name']);
                     }
                 }
@@ -823,7 +828,7 @@ class ParsingModel {
                 if($doc_b = $el->Add($arLoadArray)) {
                     echo 'Документация добавлена: ['.$doc_b.'] '.$xmlId.'<br>';
                     $id_r = $doc_b;
-                    if($this->answer) {
+                    if($this->answer && !empty($file)) {
                         unlink($file['tmp_name']);
                     }
                 }
